# Rewrite Integer#times in Ruby