--- a/gems/bundled_gems
+++ b/gems/bundled_gems
@@ -19,5 +19,5 @@ matrix          0.4.2   https://github.com/ruby/matrix
 prime           0.1.2   https://github.com/ruby/prime
 rbs             3.2.1   https://github.com/ruby/rbs
 typeprof        0.21.8  https://github.com/ruby/typeprof
-debug           1.8.0   https://github.com/ruby/debug
+debug           1.8.0   https://github.com/ruby/debug 927587afb6aac69b358b86a01f602d207053e8d2
 racc            1.7.1   https://github.com/ruby/racc