## railsbench
I benchmarked this with railsbench, which does call `Integer#times` a little (https://github.com/Shopify/yjit-bench/pull/168).

### Interpreter
It doesn't seem to have a significant impact on the interpreter.

```
before: ruby 3.3.0dev (2023-09-06T19:50:21Z master acd626a583) [x86_64-linux]
after: ruby 3.3.0dev (2023-09-06T22:03:28Z builtin-times 74e1bd7243) [x86_64-linux]

----------  -----------  ----------  ----------  ----------  -------------  ------------
bench       before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
railsbench  2018.9       0.8         2015.0      0.9         1.01           1.00
----------  -----------  ----------  ----------  ----------  -------------  ------------
```

### YJIT
It decreases `interp_return`, increases `avg_len_in_yjit`, and speeds up the benchmark a little.

```
before: ruby 3.3.0dev (2023-09-06T19:50:21Z master acd626a583) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-09-06T22:03:28Z builtin-times 74e1bd7243) +YJIT [x86_64-linux]

----------  -----------  ----------  ----------  ----------  -------------  ------------
bench       before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
railsbench  1155.4       1.4         1137.3      1.3         1.01           1.02
----------  -----------  ----------  ----------  ----------  -------------  ------------
```

## microbenchmark
### Interpreter
It slows down the interpreter a little, but it doesn't seem too significant. In fact, railsbench did not slow down.

```
$ benchmark-driver benchmark/loop_times.rb -v --chruby 'before;after'
before: ruby 3.3.0dev (2023-09-06T19:50:21Z master acd626a583) [x86_64-linux]
after: ruby 3.3.0dev (2023-09-06T22:03:28Z builtin-times 74e1bd7243) [x86_64-linux]
Calculating -------------------------------------
                         before       after
          loop_times      1.290       1.071 i/s -       1.000 times in 0.774922s 0.933629s

Comparison:
                       loop_times
              before:         1.3 i/s
               after:         1.1 i/s - 1.20x  slower
```

## YJIT
Ruby outperforms C.

```
$ benchmark-driver benchmark/loop_times.rb --chruby 'before::before --yjit-call-threshold=1;after::after --yjit-call-threshold=1' -v
before: ruby 3.3.0dev (2023-09-06T19:50:21Z master acd626a583) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-09-06T22:03:28Z builtin-times 74e1bd7243) +YJIT [x86_64-linux]
Calculating -------------------------------------
                         before       after
          loop_times      1.490       5.229 i/s -       1.000 times in 0.670956s 0.191239s

Comparison:
                       loop_times
               after:         5.2 i/s
              before:         1.5 i/s - 3.51x  slower
```