@casperisfine 2023-09-07
> In fact, railsbench did not slow down.

I wouldn't expect `Integer#times` to be called much in `railsbench`. But in general I wouldn't expect it to be called much in hotspots at all.

`Array#each` would be the big one.