@maximecb 2023-09-07
This is pretty cool. Can you run it on rubykon, mail, nbody, optcarrot and lee to make sure there's no degradation? I expect it will perform well, but it would be nice to have a bit more benchmark data.