@k0kubun 2023-09-07
> Array#each would be the big one.

Benoit thinks rewriting `Array#each` in Ruby makes it thread-unsafe https://github.com/ruby/ruby/pull/6687#discussion_r1139265240, so we're currently blocked by that.

> Can you run it on rubykon, mail, nbody, optcarrot and lee to make sure there's no degradation?

Sure

### Interpreter
```
before: ruby 3.3.0dev (2023-09-06T22:26:11Z master 54274b8c65) [x86_64-linux]
after: ruby 3.3.0dev (2023-09-06T23:26:33Z builtin-times a2bee28ecc) [x86_64-linux]

---------  -----------  ----------  ----------  ----------  -------------  ------------
bench      before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
mail       121.1        0.2         121.1       0.2         1.00           1.00
lee        926.8        0.8         938.0       0.7         0.99           0.99
nbody      93.0         1.1         85.5        0.6         1.08           1.09
optcarrot  4525.8       0.7         4544.9      0.6         1.01           1.00
rubykon    10239.2      0.5         10140.2     0.6         1.01           1.01
---------  -----------  ----------  ----------  ----------  -------------  ------------
```

### YJIT
rubykon also seems to be sped up reliably.
```
before: ruby 3.3.0dev (2023-09-06T22:26:11Z master 54274b8c65) +YJIT [x86_64-linux]
after: ruby 3.3.0dev (2023-09-06T23:26:33Z builtin-times a2bee28ecc) +YJIT [x86_64-linux]

---------  -----------  ----------  ----------  ----------  -------------  ------------
bench      before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
mail       83.2         0.2         83.6        0.2         1.00           1.00
lee        619.5        1.1         626.9       1.1         0.99           0.99
nbody      41.4         0.5         41.3        0.3         1.00           1.00
optcarrot  1236.9       0.8         1241.2      0.5         1.01           1.00
rubykon    4353.5       0.4         4253.8      0.9         1.02           1.02
---------  -----------  ----------  ----------  ----------  -------------  ------------
```